// src/types.ts
export interface PostCommentType extends Comment {
    user: {
        id: string;
        username: string;
        image: string | null;
    }
}